+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Properly close `pgx` transactions if the context used in `BeginTx` is cancelled.
 - Fix issues with generating code for relationships defined with `WhereExpr`.
 - Support `uint64` type when scanning columns as `types.Uint64`. This fixes scanning `BIGINT UNSIGNED` MySQL columns as `types.Uint64` when `interpolateParams` is enabled. (thanks @luiscleto)
+- Properly handle code generation tests for compound queries in MySQL.
 
 ## [v0.40.2] - 2025-08-16